Refactor Politician Tests
Add a base class for politician controller tests for readability and code reuse
