+++ src/test/java/com/example/demo/integration/springDoc/SpringDocAbstractTest.java
@@ -4,6 +4,7 @@ import static java.net.URI.create;
 import static org.hamcrest.CoreMatchers.containsStringIgnoringCase;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.junit.jupiter.api.Test;
 
@@ -14,6 +15,7 @@ public class SpringDocAbstractTest extends PoliticsApplicationTests {
 	@Test
 	public void testSpringDocDescriptionApi() throws Exception {
 		mvc.perform(get(create("/v3/api-docs")))
+			.andExpect(status().isOk())
 			.andExpect(jsonPath("info.description", 
 					containsStringIgnoringCase("Login with `oauth2/authorization/facebook` first to get a")));
 	}
