+++ src/test/java/com/example/demo/integration/springDoc/SpringDocVersionTest.java
@@ -4,6 +4,7 @@ import static java.net.URI.create;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.junit.jupiter.api.Test;
 
@@ -12,6 +13,7 @@ public class SpringDocVersionTest extends SpringDocAbstractTest{
 	@Test
 	public void versionShouldEqualTo100() throws Exception {
 		mvc.perform(get(create("/v3/api-docs")))
+		.andExpect(status().isOk())
 		.andExpect(jsonPath("info.version", 
 				equalTo("1.0.0")));
 	}
