+++ src/test/java/com/example/demo/unit/controller/PoliticianControllerTest.java
@@ -7,47 +7,15 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.math.BigDecimal;
-
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import com.example.demo.controller.PoliticianController;
+import com.example.demo.baseClasses.AbstractPoliticianControllerTest;
 import com.example.demo.dto.PoliticianDTO;
-import com.example.demo.dtoRequest.AddPoliticianDTORequest;
 import com.example.demo.dtomapper.PoliticiansDtoMapper;
-import com.example.demo.model.averageCalculator.LowSatisfactionAverageCalculator;
-import com.example.demo.model.entities.Politicians;
-import com.example.demo.model.entities.Rating;
-import com.example.demo.service.PoliticiansService;
 
-@ExtendWith(SpringExtension.class)
-public class PoliticianControllerTest {
+public class PoliticianControllerTest extends AbstractPoliticianControllerTest {
 
-	@Mock
-	public PoliticiansService service;
-	public PoliticianController controller;
-	public Politicians politician;
-	public AddPoliticianDTORequest politicianDTORequest;
-	
-	@BeforeEach
-	public void setUp() {
-		controller = new PoliticianController(service);
-		
-		politician =  new Politicians();
-		politician.setRating(new Rating(9.67D, 9.67D, new LowSatisfactionAverageCalculator(9.07D, 0D)));
-		politician.setFirstName("Mirriam");
-		politician.setLastName("Defensor");
-		
-		politicianDTORequest = new AddPoliticianDTORequest
-		("Mirriam", 
-		"Defensor", 
-		BigDecimal.valueOf(9.67D));
-	}
 	
 	@Test
 	public void shouldCallSaveFromServiceWhenSaved() {
